requirements.txt
pynetbox>=7.5.0
pydantic>=2.12.0
pyyaml
typer[all]
python-dotenv
rich